added BehaviourSubject to maintain authState on client side logout, login redirects and fixes

+++ src/app/todos/todo-list/todo-list.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { PageEvent } from '@angular/material/paginator';
 import { Router } from '@angular/router';
-import { AlertService } from 'src/service/alert.service';
+import { Subscription } from 'rxjs';
 
+import { AlertService } from 'src/service/alert.service';
+import { AuthService } from 'src/service/auth.service';
 import { TodoService } from 'src/service/todo.service';
 import { Todo } from './../../../data-models/todo.model';
 
@@ -11,9 +13,10 @@ import { Todo } from './../../../data-models/todo.model';
   templateUrl: './todo-list.component.html',
   styleUrls: ['./todo-list.component.css']
 })
-export class TodoListComponent implements OnInit {
+export class TodoListComponent implements OnInit, OnDestroy {
 
   todos: Todo[] = [];
+  todosSub: Subscription;
   isLoading = false;
 
   length = 0; // totalTodos
@@ -21,15 +24,18 @@ export class TodoListComponent implements OnInit {
   pageIndex = 0; // Current Page (starting with 0)
   pageSizeOptions = [1, 2, 5, 10];
 
-  constructor(private todoService: TodoService, private router: Router, private alertService: AlertService) { }
+  authState: boolean;
+
+  constructor(private todoService: TodoService, private router: Router, private alertService: AlertService, private authService: AuthService) { }
 
   ngOnInit(): void {
     this.getTodos();
+    this.authService.authStateListener().subscribe((data) => this.authState = data);
   }
 
   getTodos() {
     this.isLoading = true;
-    this.todoService.getTodos(this.pageSize, this.pageIndex).subscribe((res) => {
+    this.todosSub = this.todoService.getTodos(this.pageSize, this.pageIndex).subscribe((res) => {
       // console.log(res);
       this.isLoading = false;
       this.todos = res.todos;
@@ -57,12 +63,19 @@ export class TodoListComponent implements OnInit {
   }
 
   deleteAllTodos() {
+    if (!window.confirm("Are you sure you want to delete all your todos?")) return;
     this.todoService.deleteAllTodos()
     .subscribe((res) => {
       this.getTodos();
       console.log(res.message);
       this.alertService.openSnackBar("Deleted all your todos Successfully");
     });
+  }
+
+  onExpand() { }
+
+  ngOnDestroy(): void {
+    this.todosSub.unsubscribe();
   }
 
 }
